basic events structure

See merge request CS309SPRING2019/team_7!4

+++ Backend/309_Server/src/main/java/callyourmom/service/EventService.java
@@ -18,12 +18,11 @@ public class EventService {
 		return eventRepository.findAll();
 	}
 
-	public Event get(Long id) {
+	public Event get(Integer id) {
 		return eventRepository.getOne(id);
 	}
 
 	public Event create(Event event) {
 		return eventRepository.save(event);
 	}
-
 }
